Port of Qt to VxWorks
This makes Qt work on VxWorks 6.6+ in native (kernel) mode.
 * compiles with the WindRiver GNU toolchain (Linux only)
 * works with QWS (tested with the VNC driver only)
 * tested on PPC hardware and the x86 VxWorks simulator
 * no q3support, no phonon, no webkit
 * no QSharedMemory, no QSystemSemaphore, no QProcess
 * only one QApplication instance (flat address space)
 * filesystem support depends heavily on the quality of the native driver
 * QLibrary is just a dummy to make plugins work at all
 * qmake transparently creates VxWorks munching rules for static ctors
 * made auto-test cope with missing OS features

A special note regarding the Q_FOREACH patch for dcc:
when calling foreach(a,c) with c being a function returning a container,
the compiler would generate 5 references to some labels (.LXXXX), which
are not there (so the linker complains in the end).
Seems like dcc doesn't really like the 'true ? 0 : <function call to get type>'
statement

Reviewed-By: Harald Fernengel