Add LIBS_PRIVATE to qmake.
The difference between LIBS and LIBS_PRIVATE is that private libraries
are those that are not part of the public interface of a library. For
example, if you're writing a Qt application and link to QtGui, you
definitely need the development files for QtCore, but not necessarily
for Glib and GThread, or maybe even X11.

The private libraries are necessary only in static builds, so the
information should still be published in .prl and pkg-config files.

Reviewed-By: Marius Storm-Olsen