implement event propagation for touch events
behavior is similar to that of QGraphicsScene mouse events: the first
touch event (the TouchBegin) is propagated to all parents. if a widget
accepts the event, it will receive all other touch events (the
TouchUpdate and TouchEnd events). If no widget accepts the TouchBegin,
then we will fallback to normal mouse events (TBD).