Better look of toolbar items on Leopard.
Toolbar text items have a embossing on Leopard, so we should do that.
However, they have a different look when "pressed" depending on whether
or not they have an icons. We are very close to this effect (depending
on how transparency rounding goes). Steps have also been taken so that
the Tiger look is preserved. This is step one in several of getting the
toolbars to look better on Leopard and friends.

Task-number: 248006
Reviewed-by: Jens Bache-Wiig