Fix assertion failure in QCompleter::setCompletionPrefix().
Calling setCompletionPrefix() on a QCompleter from a slot connected
to the editingFinished() signal of the corresponding QLineEdit
could in some cases alter the internal state of the completer in
such a way that an assertion would fail. The fix prevents the
asserting code from being called in this particular state.

Reviewed-by: janarve
Task-number: 246056