Make sure image caching works with Qt3Support
Store whether qt was built with qt3support enabled in .qmake.cache. We
need this info to decide if QT3_SUPPORT should be defined when we
include qimage_p.h. Otherwise we get data corruption issues since the
size of struct in Qt is larger than the one we include in libqdirectfb.

Reviewed-by: Donald <qt-info@nokia.com>