Fix a few issues introduced by Ben's last patch.
Test if we're not past the end of the array before we attempt to read it.

Dates without timezone information are taken to be UTC, so merge the code.

If the date parsing failed, don't add a malformed cookie. Better to
have no cookie than to have it for past its expiration date.

Also update the autotests, since we now can cope with some of the
invalid cookies seen on the net.

Signed-off-by: Peter Hartmann <peter.hartmann@trolltech.com>