Fix regression in qscriptvalue_cast (primitive-->complex type cast)
qscriptvalue_cast implementation was changed from 4.4 to 4.5 because
of the introduction of QScriptValue constructors that don't take an
engine pointer. However, when the old constructors are used, the
behavior of qscriptvalue_cast should be as before, which this patch
ensures. In short: If we have an engine pointer, use it.

Task-number: 248802
Reviewed-by: Ariya Hidayat