Fix a bug where lineedits in a second page of a stack widget would not get key events.
We need to be more explicit in setting the first responder, and a bit
more agressing about rejecting it when called with no responder.

It seems that Cocoa has a tendency to reset the first responder when
lots of widgets get hidden or shown. During this it will call
"makeFirstResponder" on the window with a nil responder. Doing this will
reset our what Cocoa thinks is the focus widget, but Qt will still show
a focus widget. The way to solve it is to reject the make first
responder if the responder is nil. I'm not sure if there will be far
reaching implications for this, but it seems to be doing the right thing
at the moment.

We also need to share this code between QCocoaWindow and QCocoaPanel.
Thanks to the dynamic nature of objective-C we have to have a copy of
the code in both places. It's unfortunate.

Finally, it's also important to have the QWidget let Cocoa know it has
focus after it has been created. So, make sure that is in sync.

Task-number: 249296
Reviewed-by: Prasanth Ullattil