don't include uic in non-gui configurations
uic is set in default_pre. this leads to useless -I compiler flags when
no ui files are used in fact, so it would be nice to get rid of it in
cases where it is not used. for backwards compat, don't remove it from
the _pre, but add some magic to _post to remove it again if QT does not
contain gui. of course, we need a force_uic CONFIG to enable an
exception for QtGui itself ...