Fix crashes in D-Bus due to race conditions.
This happens in heavily threaded applications: libdbus-1 asks us to add
and remove timers (and watches, but those less often) from any thread.
Since QObject's model requires that timers and QSocketDescriptors be
added only in the object's own thread, we have to sync back to the main
thread.

However, since D-Bus keeps adding and removing the timers, we have to
keep a queue of timers to be added and that list is accessed by any
thread. Make sure that list is only accessed in locked conditions.

Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>