String-to-number conversion functions should ignore trailing whitespaces.
According to our documentation we should ignore leading and trailing
whitespaces when converting a string to number with QLocale::toInt and
similar functions. However that didn't work for some locales - for
those ones that declare groupseparator as 0xa0 (which looks similar to
space) since we provide a workaround to accept space as a group
separator for those locales. And since the workaround was there for a
long time it doesn't make sense to change the behavior and the fix is
to explicitely remove leading and trailing whitespaces before doing
any conversion.

Reviewed-by: mariusSO