Don't send QFileOpenEvents for items on the command-line
Cocoa actually has a nice feature that if you pass arguments on the
command-line, Cocoa will pass those along as Open events later. This is
probably how we should have handled things inside of Qt as it would have
unified the file opening code. Unfortunately, we can't turn back time on
this, so we need to prevent it because people probably aren't expecting
it (i.e., they expect to do the parsing themselves, and not to get
events later). This also means that we can send the event immediately
instead of posting it, because the race that we had before no longer
exists. We only do this check during launch time because that's the only
time we may get bitten by it (people usually only parse the arguments
once).

Someday, people may actually WANT this functionality though. When that
comes along, we should make it an application attribute.

Task-number: 249553
Reviewed-by: Richard Moe Gustavsen