Get the qt7 phonon plugin to minimally compile on 10.6
We need to protect some overridden type and that is done in
qmacdefines_mac.h, so just include that. To top this off, a
function has changed that was introduced in 10.5. In order to
use it properly, we would need to do lots of version checking
and shuffling between types and it seems like more trouble that
it's worth at the moment. We may have to do that in the future,
but we'll deal with that then.