Speed up selectAll in QTreeView with spans
When we had spans, in order to have the selection rect, we took the slow
path in all cases, even if the header section did not move.

Now, take the fast path if the sections did not move.

This requires few adjustement to the code: we cannot call visualRect
anymore since the index may be within a span.

Task-number: 244651
Reviewed-by: Marius Bugge Monsen