Work on getting unified toolbars to look more native.
Basically we try to get the toggled look correct and we've shrunk the
size of the toolbar by a good 10 pixels. We still look a bit "off" for
toggled on Tiger, but frankely that look is a bit odd.

We are a bit taller than the pure Cocoa toolbar (2 px), but given that
we are embedding our QToolbar, that's probably the best we can do.

All-in-all, it looks much better.