Delete temporary file if copy was unsuccessful
This is an untested bugfix -- comes solely from reading the code.

In QFile::copy's fallback implementation a (temporary) file is created
for block copying from the source file. When Qt is built without
temporary file support this doesn't seem to be deleted in case of an
error while block copying or renaming to the final destination.

Reviewed-by: mariusSO