Fixing some CoreText errors.
Apparently CoreText is too good for us and always gives us fractional
information. Which would be food for design metrics, but not good for
drawing to the screen. The upshot of it is that we do the ceiling of
every value we get. This is as stop gap solution for the moment. Things
in Creator should look better now.

Also added the ability to disable kerning.

Reviewed-by: Simon Hausmann