Fix focus policy propagation bug in QCompleter.
Commit 4a00810cc394b7da99ddb2d13eb045352ede25d3 ensures
that setting the focus policy of a widget applies to its
proxy (if any) as well. This propagation effect is however
unfortunate for a QCompleter in PopupCompletion mode.
The client widget (typically a QLineEdit or a QTextEdit)
acts as the focus proxy to the popup menu. Internally,
the completer sets the focus policy of the popup menu to
NoFocus, but this should not affect the focus policy of
the editor.

Reviewed-by: janarve
Task-number: 250064