Fix QNetworkReply's automatic pushing of a cache when the cache is more than 64k.
This issue was found by Warwick: the copyReadyRead() slot is called
only once as a result of readyRead(), but since there's no more data
to be received (the source QIODevice is a QBuffer), we'll never get a
readyRead() again.

So, if we don't have a limited buffer size, we should read
everything. This applies to QFile as well.

The side-effect is that we cause the entire QFile to be loaded to
memory, so if you had a 2 GB cache entry, you'll probably run out of
memory.

Future optimisations:
 - don't memcpy the data from a QBuffer into another buffer
 - find a way to avoid loading the entire contents of the file
 (probably not possible with the default QNetworkDiskCache since that
 may compress the data, so we won't know the size and thus can't fake
 finished())

Reviewed-by: Warwick Allison