Fix up color matching on 64-bit Cocoa.
It seems that GetThemeTextColor is not available at all for 64-bit
versions of the operating system and there is no good equivalent.
We were using a technique where we would create a small pixmap and then
use HIThemeSetTextFillColor() with the proper enum. However, this
function will silently set a black color if it doesn't know about the
enum. So, we get black when we really shouldn't. It also seems to
produce colors that are a bit different from the Carbon function. So, in
the meantime, just hard code the values in. It might break if you are
hacking the resources in Mac OS X, but we'll have to live with that
until we get a real function.

Task-number: 248918
Reviewed-by: Prasanth Ullattil
(cherry picked from commit c350892f55a1d8633be54f09e6e4596f4687bc89)