Fix moc autotest failures
We have to pass the correct include paths when testing the multiple
inheritance warning, because we want moc to find the QtGui and QObject
headers.

In addition we have to adjust the line numbers and include QObject in
the qobject test to avoid a warning about the unknown QObject
superclass.

Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>