BT: Adjust the colliding mice example to work with coalesced updates.
It seems that Cocoa is much more strict about coalesced updates than
Carbon ever was. The upshot of this is that some examples that "worked"
after a fashion in Carbon, do not exhibit good frame rates with Cocoa.

The reason why is that apparently Cocoa will decide to flush to the
screen every time a timer fires. If you have a lot of timers that are
all dependent on doing on update to the screen, you will get undesirable
effects.

Thankfully, it is possible to adjust the examples to follow best
practices and get a good result. So, we now only do the animation once
using QGraphicsScene::advance(). We are also able to make the mice less
heavy (no QObject subclass). I've updated the docs and someone on the
doc team has kindly volunteered to go through them.

Reviewed-by: Andreas