BT: Fix regression when tooltips dissappear suddenly in Unified toolbar
QWidget::childAt() makes some assumptions about its children (they are
all contained in its geometry).  This does not hold up when using the
unified toolbar because the toolbar ends up in the "non-client" area.
So, when dispatching an enter/leave event in tooltip show, we end up
dispatching to the wrong widgets and that results in the tooltip
cleverly thinking that it needs to hide itself because we've left the
widget that needs the tooltip. I've special cased this by just having a
"native" mapFromParent() that is only called for on the mac, though
there is nothing that is limiting this from being called on other
platfroms.

Also QWidget::mapFromParent() probably needs to be looked at at some
point.

Task-number: 248048
Reviewed-by: Richard Moe Gustavsen
(cherry picked from commit 640f2c732c6fd76866cd7e601a9140dbe7849e6f)