make shadow builds with default moc/ui dirs work again
append the source dir to the include path, but only after the dirs with
the generated files.
this seems to have worked before only accidentally: the unqualified
default dirs were expanded to the source dir instead of the build dir,
but the build dir is added implicitly by default, so things magically
worked. now that we qualify the moc/ui dirs, projects relying on the
strange side effect suddenly break.
we should probably add the source dir to the include path by default,
but this coupling to uic/moc is closer to the historical behavior and
thus should be safer.

Reviewed-by: mariusSO