Optimize bytesPerLine further
If we're asking for the stride it's very likely the next thing we'll do
is ask for the bits() so there's no good reason to unlock it again.

In the raster buffer case memory() will be called just before
bytesPerLine() so the code won't be hit but it's still the right thing
to do.

Reviewed-by: TrustMe