Bt: Fix regression in the Embedded dialogs example
We had to revert an earlier fix since it obviously did
not work correctly. However since we do not really need to propagate the
palette on the viewContainer _before_ it is created, we can simply avoid
the issue alltogether as it would happen because we implicitly added
a child widget during the polish of the combo box.

Reviewed-by: nrc
(cherry picked from commit 35c26d696cbff269d551c012a212c09692dd6f6b)