BT: Revert "Fixes:    Remove warnings in the embeddeddialogs demo."
This reverts commit 7d24c0af8e52687b4e5735950495bb25947ca9f0.

This change removed the accidental recursion leading to console
warnings on Mac OS X by changing the way the scene event filter
is registered for popups. But it also broke the behavior of the
demo with regards to popup handling, in that the custom proxy
no longer can detect popups that close, so dialogs stay zoomed
in after the cursor has left. So before, the demo worked but
printed warnings to the console on Mac OS X, and after, the warnings
were gone but the demo behavior broke on all platforms...

The change also accidentally removed the cache mode from the popup;
I don't think this was intentional.

Reviewed-by: Alexis
(cherry picked from commit ccaf16278f4fd422458d01b80579952846e6cf38)