Revert "Bt: Fix regression in the Embedded dialogs example"
This reverts commit 35c26d696cbff269d551c012a212c09692dd6f6b.

The change to QComboBox introduces a behavior change; whereas before
the view container would always get its palette set as a response
to QEvent::PaletteChange, it would now miss this event and rely on
regular palette propagation to get the right contens. The difference
in behavior is that QWidget::setPalette() also resolves the palette
mask, and after 35c26d69 this would no longer happen.

The bug in the embedded dialogs demo is caused by the embedded
dialogs demo. See upcoming commit.

Reviewed-by: Alexis
Reviewed-by: Jens Bache-Wiig <jbache@trolltech.com>