BT: Fix the embedded dialogs demo on Mac OS X and Motif style
The CustomProxy class installs an event filter on its child to detect
whether it is shadowed by a popup or not. The problem is it does this
regardless of whether it currently has a scene assigned or not. Styles
that assign palettes, or otherwise cause side effects when assigned
to a QGraphicsProxyWidget, will cause the demo to print warnings to the
console and fail to install the event filter. The reason for the failure
to install the filter is that QGraphicsItem only allows scene event
filters to be installed between items that are in the same scene.

So, depending on the style, you either get an ItemSceneHasChanged or an
ItemChildAddedChange first. The demo must account for this, and install
its filter only when the items are guaranteed to be in the scene already.

Reviewed-by: Alexis