Take out the DSCAPS_TRIPLE when not blitting
When this code was written there was never a case when we would not blit
our flips and hence the entire if (!(flipFlags & DSFLIP_BLIT)) business
seems not to have been tested anyway. Since I don't see the point of
this I am taking it out. Might enable us to actually create the primary
surface in video memory more often.

Reviewed-by: TrustMe