Don't remove the cache file if you have successfully added it.
This regression probably happened because of the fix to task 244485
(see 8d500381), which made QFile follow a rename. This means that
QTemporaryFile removes its target when it is deleted.

This fixes a number of caching autotests that are failing.

Reviewed-by: Markus Goetz
BT: yes
(cherry picked from commit 99ed68cda10355dba5c7a3e2466ee45b59dbb555)