Fixes the composition demo for Mac/Cocoa in GL mode.
QGLPixelBuffer::generateDynamicTexture() will bind the texture to the
pbuffer regardless. Why this works on Carbon is a mystery, but if we're
to follow our own docs, we should NOT bind the texture to the pbuffer
by default. An explicit call to ::bindToDynamicTexture() is required for
that.

Task-number: 250664
Reviewed-by: Samuel
BT: yes
(cherry picked from commit e8877160e332d12958cf700a8a607c36e6c426e1)