BT: Fix a crash on certain theme changes in GNOME
We did not update initGtkWidgets on theme changes which could
lead to crashes as the contents of the widget hash depend on
certain properties in the style. A simple way to reproduce this would
be to change between the redmond gtk theme and cleanlooks.

Task-number: 251115
Reviewed-by: paul
(cherry picked from commit a55f477b2a40b89e690fe6b61509cf665e96243b)