Make subpixel antialiasing of text work on Mac OS X under carbon
Implement the alphaRGBMapForGlyph function and make use of it in
the raster engine. The gamma correction is currently hardcoded to 2.0
which looks very good on two separate displays (iMac and MacBook Pro).
Ideally this would be picked from the system settings or computed
dynamically, but as long as 2.0 works, we'll stick to that.

Reviewed-by: Samuel