Add a warning for trying to connectToHost() when disconnecting.
Right now, we print this warning if connectToHost() is called when
we're in ConnectingState (waiting for remote to SYN,ACK) or in
ConnectedState. This also means connectToHost() allows interruption of
the HostLookupState cleanly (I think).

But if you called connectToHost() right after disconnectFromHost() and
there were data to write, the connectToHost() could discard the
pending outgoing data and reconnect. Or not, depending on whether the
DNS resolution ended first. In other words, race condition.

Reviewed-by: Peter Hartmann