Fix QPrinter::setPrinterName()/QPrinter::printerName() on the Mac.
QPrinter::printerName() used to return the CUPS printer queue name instead
of the user-set printer name, which was wrong. In addition, if trying to
set the printer name to a non-valid printer, the default printer was still
used internally, which is not correct. As sonn as an invalid name is set,
the printer now enters an invalid state, which it can only get out of
by setting a valid printer name.

Task-number: 199271
Reviewed-by: Gunnar