BT:Stop widgets disappearing in unified when clicking the toolbar button
The comment in the code pretty much says it all, so I'll repeat it:

This is a workaround for NSToolbar. When a widget is hidden by
clicking the toolbar button, Cocoa reparents the widgets to another
window (but Qt doesn't know about it).  When we start showing them,
it reparents back, but at this point it's window is nil, but the
window it's being brought into (the Qt one) is for sure created.
This stops the hierarchy moving under our feet.

Task-number: 251294
Reviewed-by: Prasanth Ullattil
(cherry picked from commit 139265031939d676f049edd2c91918040e0659b4)