Fix two bugs in Graphics View's mouse grab logics + add autotests.
Change fd4d94df4eb16e9e54f61dd8ee45914e24832ae9 introduced mouse grab
and keyboard grab support to Graphics View (this was as part of 4.4).
The change was missing a large set of autotests that I wrote to test
the behavior of these features. As part of task 245317, which involves
investigating modality in Graphics View, I figured it would be a good
idea to start off but reconstructing these autotests. So this change
is mainly about adding autotests for mouse grabbing. And of course,
as it always is, I found two bugs while writing these tests.

1) Calling QGraphicsItem::grabMouse() while the item is holding the
   implicit mouse grab will now upgrade the implicit grab to an
   explicit grab.

2) Adding a popup to the scene will automatically grab the mouse.
   Before, the popup would get the grab on show(), but if it was
   already visible when added to the scene it would not gain the
   grab.

Task-number: 245317
Reviewed-by: jasplin