BT: Make native QFontDialog work when using show and changing windows
Cocoa's font manager uses "first responder" which is a great idea, but
breaks as soon as we change windows. Thankfully we can just set the
target and we are OK. An upshot is that we don't need the delegate, but
I'm not going to push my luck on that.

Reviewed-by: Richard Moe Gustavsen
(cherry picked from commit 7c815bc2d4ea922d75f49f7dc29c81362ce5210e)