Fix tabbar issues with Oxygen style
Oxygen style was looking a bit broken when using 4.5 features such as
close buttons and icons. This was because it re-implements the
tabbartablabel primitive we needed to ensure spacing for the buttons.
We had to modify the way we draw tabs slightly to make it more backward
compatible so that oxygen get the same region as before and we ensure
space for it in the CE_TabBarTab case instead. Note that the drawback
to this approach is that the focus rect needs to be moved out of the
tablabel and into the CE_TabBarTab case so it covers the entire tab
which can potentially cause problems if people depend on the label to
draw the focus rect. However the new case is also more consistent with
how buttons draw their focus.

Task-number: 251635
Reviewed-by: nrc