Missing debug .rc file with a clean shadow build
When generating Windows Makefiles, qmake writes out a .rc file for each
of debug and release (unless you've limited to just one build type).
When doing a clean shadow build, the first .rc file is written into a
directory that does not exist but the code was not handling the error
case. The fix does 2 things.
1) Attempt to create the destination directory if we can't write the file.
2) Die with an error if we still can't write the file after doing #1.

Reviewed-by: Marius Storm-Olsen