BT: Revert "Fixes:    Don't fill the pixmap because we will copy the cache into it."
This reverts commit 0985805ab3c7de5b15c115a98afb15944b6d93b9.

The Elastic Nodes example shows a black background for the cache pixmap
if the call to fill isn't there. Alexis and I both agree that it makes
no sense that we fill the pixmap with transparent pixels instead of
just drawing with QPainter::CompositionMode_Source, but since this
regression was introduced after 4.5.0 was released we must revert before
we investigate further.

Reviewed-by: Alexis