Deprecate qt_mac_set_show_menubar for a public cross-platform API.
I'm tired of these "hidden" functions. We have an
AA_MacPluginApplication, but sometimes you may have a legitimate reason
for setting this outside of "plugin applications." In the footsteps of
the menu icon attribute, the attribute is the main leader, but menubars
can disable/enable this locally the new QMenuBar::setNativeMenuBar()
property. Otherwise, the menubars take their que from the application
attribute.

This also works for Windows CE. So, there is a bit on convergence as
well.

Task-number: 236757